dev-apcus : Modificacion de unos tipos de datos.

+++ app-customer/src/main/java/com/petproject/appcustomer/domain/ports/in/RetrievePetUseCase.java
@@ -5,7 +5,7 @@ import com.petproject.appcustomer.domain.models.in.PetEntity;
 import java.util.List;
 
 public interface RetrievePetUseCase {
-    PetEntity getPet(Long id);
+    PetEntity getPet(Integer id);
 
     List<PetEntity> getAllPet();
 }
