+++ app-customer/src/main/java/com/petproject/appcustomer/domain/ports/in/UpdatePetUseCase.java
@@ -3,5 +3,5 @@ package com.petproject.appcustomer.domain.ports.in;
 import com.petproject.appcustomer.domain.models.in.PetEntity;
 
 public interface UpdatePetUseCase {
-    PetEntity updatePet(Long id, PetEntity petEntity);
+    PetEntity updatePet(Integer id, PetEntity petEntity);
 }
